Delete moved to Options Package
