+++ src/main/java/com/pelensky/contactmanager/Commands.java
@@ -1,17 +1,17 @@
 package com.pelensky.contactmanager;
 
-class Commands {
-    IO io;
+public class Commands {
+    private IO io;
     ContactList contactList;
     private ManipulateContacts manipulateContacts;
 
-    Commands(IO io, ContactList contactList, ManipulateContacts manipulateContacts) {
+    public Commands(IO io, ContactList contactList, ManipulateContacts manipulateContacts) {
         this.io = io;
         this.contactList = contactList;
         this.manipulateContacts = manipulateContacts;
     }
 
-    int selectContactTo(String text) {
+    protected int selectContactTo(String text) {
         io.displayText(
                 "Which contact would you like to " + text + "?"
                         + System.lineSeparator()
