+++ src/main/java/com/pelensky/contactmanager/Options/DefaultOption.java
@@ -1,10 +1,13 @@
-package com.pelensky.contactmanager;
+package com.pelensky.contactmanager.Options;
+
+import com.pelensky.contactmanager.IO;
+import com.pelensky.contactmanager.Option;
 
 public class DefaultOption implements Option {
 
     private IO io;
 
-    DefaultOption(IO io){
+    public DefaultOption(IO io){
         this.io = io;
     }
